requirements.txt
pandas==3.0.2
pyarrow==24.0.0
snowflake-connector-python==4.4.0
